refactor: split schema init by provisioner and fix db9 PG DDL
