+++ pkg/tenant/schema_starter.go
@@ -1,0 +1,93 @@
+package tenant
+
+import (
+	"database/sql"
+	"fmt"
+
+	_ "github.com/go-sql-driver/mysql"
+)
+
+func initStarterSchema(dsn string) error {
+	if hasMultiStatements(dsn) {
+		return fmt.Errorf("multiStatements is not allowed")
+	}
+	db, err := sql.Open("mysql", dsn)
+	if err != nil {
+		return err
+	}
+	defer func() { _ = db.Close() }()
+	if err := db.Ping(); err != nil {
+		return err
+	}
+	if !isTiDBCluster(db) {
+		return fmt.Errorf("provider requires TiDB capabilities (FTS/VECTOR)")
+	}
+
+	stmts := []string{
+		`CREATE TABLE IF NOT EXISTS file_nodes (
+			node_id      VARCHAR(64) PRIMARY KEY,
+			path         VARCHAR(512) NOT NULL,
+			parent_path  VARCHAR(512) NOT NULL,
+			name         VARCHAR(255) NOT NULL,
+			is_directory BOOLEAN NOT NULL DEFAULT FALSE,
+			file_id      VARCHAR(64),
+			created_at   DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3)
+		)`,
+		`CREATE UNIQUE INDEX idx_path ON file_nodes(path)`,
+		`CREATE INDEX idx_parent ON file_nodes(parent_path)`,
+		`CREATE INDEX idx_file_id ON file_nodes(file_id)`,
+		`CREATE TABLE IF NOT EXISTS files (
+			file_id         VARCHAR(64) PRIMARY KEY,
+			storage_type    VARCHAR(32) NOT NULL,
+			storage_ref     TEXT NOT NULL,
+			content_blob    LONGBLOB,
+			content_type    VARCHAR(255),
+			size_bytes      BIGINT NOT NULL DEFAULT 0,
+			checksum_sha256 VARCHAR(128),
+			revision        BIGINT NOT NULL DEFAULT 1,
+			status          VARCHAR(32) NOT NULL DEFAULT 'PENDING',
+			source_id       VARCHAR(255),
+			content_text    LONGTEXT,
+			embedding       VECTOR(1024) GENERATED ALWAYS AS (EMBED_TEXT('` + autoEmbedModel + `', content_text, '{"dimensions": 1024}')) STORED,
+			created_at      DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
+			confirmed_at    DATETIME(3),
+			expires_at      DATETIME(3)
+		)`,
+		`CREATE INDEX idx_status ON files(status, created_at)`,
+		`ALTER TABLE files
+			ADD FULLTEXT INDEX idx_fts_content(content_text)
+			WITH PARSER MULTILINGUAL
+			ADD_COLUMNAR_REPLICA_ON_DEMAND`,
+		`ALTER TABLE files
+			ADD VECTOR INDEX idx_files_cosine((VEC_COSINE_DISTANCE(embedding)))
+			ADD_COLUMNAR_REPLICA_ON_DEMAND`,
+		`CREATE TABLE IF NOT EXISTS file_tags (
+			file_id   VARCHAR(64) NOT NULL,
+			tag_key   VARCHAR(255) NOT NULL,
+			tag_value VARCHAR(255) NOT NULL DEFAULT '',
+			PRIMARY KEY (file_id, tag_key)
+		)`,
+		`CREATE INDEX idx_kv ON file_tags(tag_key, tag_value)`,
+		`CREATE TABLE IF NOT EXISTS uploads (
+			upload_id          VARCHAR(64) PRIMARY KEY,
+			file_id            VARCHAR(64) NOT NULL,
+			target_path        VARCHAR(512) NOT NULL,
+			s3_upload_id       VARCHAR(255) NOT NULL,
+			s3_key             VARCHAR(2048) NOT NULL,
+			total_size         BIGINT NOT NULL,
+			part_size          BIGINT NOT NULL,
+			parts_total        INT NOT NULL,
+			status             VARCHAR(32) NOT NULL DEFAULT 'UPLOADING',
+			fingerprint_sha256 VARCHAR(128),
+			idempotency_key    VARCHAR(255),
+			created_at         DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
+			updated_at         DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3) ON UPDATE CURRENT_TIMESTAMP(3),
+			expires_at         DATETIME(3) NOT NULL,
+			active_target_path VARCHAR(512) AS (CASE WHEN status = 'UPLOADING' THEN target_path ELSE NULL END) STORED
+		)`,
+		`CREATE INDEX idx_upload_path ON uploads(target_path, status)`,
+		`CREATE UNIQUE INDEX idx_idempotency ON uploads(idempotency_key)`,
+	}
+
+	return execSchemaStatements(db, stmts)
+}
